+++ CHANGELOG.md
@@ -6,10 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [ 1.0.0 ] - 2022-02-18
-### Addded
-- Introduced IConfigurationStore
-### Changed
+### Added
+- Introduced IEnumConfigurationStore and sample implementation PropertiesEnumConfigurationStore
+### Fixed
 - Improved error messages.
+- Several small bugfixes.
 
 ## [ 0.9.13 ] - 2021-12-05
 ### Changed